added the minionRouter file to handle all minion related request
